getting the change
